+++ setup_ml_env.sh
@@ -1,9 +1,13 @@
 #!/bin/bash +x
 module load conda
 export ML_ENV=$HOME/PolarisAT/pyenvs/ml_workload
-conda create  -p $ML_ENV --clone  /soft/datascience/conda/2022-09-08/mconda3/
-pip install git+https://github.com/hariharan-devarajan/dlio-profiler.git
-conda activate $ML_ENV
-
+if [[ -e $ML_ENV ]]; then
+    conda activate $ML_ENV
+else
+    conda create  -p $ML_ENV --clone  /soft/datascience/conda/2022-09-08/mconda3/
+    conda activate $ML_ENV
+    pip install git+https://github.com/hariharan-devarajan/dlio-profiler.git
+fi
+# INSTALL OTHER MISSING FILES
